Add logic, + vibration on success

+++ Rest WatchKit Extension/HeartRate.swift
@@ -11,6 +11,6 @@ import Foundation
 struct HeartRate {
 
     let timestamp: Date
-    let bpm: Double
+    let bpm: Int
 
 }
